Adding Tag System and introducing the Mover - Component
The Mover Component will be the only way to move, rotate, or scale a
GameObject.

This is good because the engine can now "detect" changes in position,
rotation, scale.

Could be useful for optimizing the Chunk - system so that a GameObject
will only be checked whether it is in a chunk if it moved.



(I kinda messed up the order of the Transform components...

Before: Transform(position, scale, rotation)
Now:		Transform(position, rotation, scale)

and I really dont know why I did that...)